fix(card): update package declaration on moved Card/Deck tests
Git's rename detection picked up the file move but kept the original
`package database.poker` line — JUnit was loading them under the old
package, leaving the new database.card directory un-tested.
